refactor: share the damage-apply buttons across attack cards
怪異カードにも武器カードと同じ〔ダメージを適用〕〔軽減して適用〕を出し、
ハンドラの実体を applications/attack-card.ts に1つだけ置いた。
怪異カードのモデルも AttackCardModel を継承する形に寄せてある。

これで既存のバグが1件直る。applyKaiDamage はダイアログが返した armor
（部位条件で防具を外したときの上書き）を捨てていて、怪異の攻撃を受ける
共鳴者の防具指定が届いていなかった。

ボタン列のCSSは2枚で同じ並びになるので、chat 層の中でセレクタを並べる形
（css/chat/attack-card.css）に寄せた。
