+++ css/chat/attack-card.css
@@ -1,0 +1,25 @@
+/* 攻撃カード（武器・怪異）で共通のボタン列。
+
+   どちらも chat 層に閉じているので部品は作らず、セレクタを並べる
+   （docs/ui-design.md「重複をどこに寄せるか」）。ここに置くのは、
+   片方のカードのファイルにもう片方のクラス名を書かせないため */
+
+@layer chat {
+  /* ボタンは縦に積む。チャットの幅は狭く、横並びにすると文字が潰れる */
+  .em-weapon-card__buttons,
+  .em-kai-attack-card__buttons {
+    display: flex;
+    flex-direction: column;
+    gap: var(--spacer-4);
+  }
+
+  /* 地・枠・角丸・ホバーは本体の --button-* が持っているので塗り直さない。
+     アイコンと文字の間隔だけ足す */
+  .em-weapon-card__buttons button,
+  .em-kai-attack-card__buttons button {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    gap: var(--spacer-4);
+  }
+}
