+++ css/chat/kai-attack-card.css
@@ -27,10 +27,8 @@
       margin: 0 0 var(--spacer-2);
     }
 
+    /* 並びは css/chat/attack-card.css が武器カードと共通で持つ。ここは間隔だけ */
     .em-kai-attack-card__buttons {
-      display: flex;
-      flex-wrap: wrap;
-      gap: var(--spacer-4);
       margin-top: var(--spacer-8);
     }
   }
