+++ module/data/messages/kai-attack-card.ts
@@ -1,17 +1,15 @@
 import type { CardActions } from "../../utils/chat-card";
-import { type CardIdentity, ChatCardModel } from "./card-model";
+import { AttackCardModel, type AttackProgress } from "./attack-card";
+import type { CardIdentity } from "./card-model";
 
 const { BooleanField, DocumentUUIDField, NumberField, StringField } = foundry.data.fields;
 
 /** カードの描画・保存に要る状態。スキーマと同じ形 */
-export type KaiAttackCardState = {
+export type KaiAttackCardState = AttackProgress & {
   attackName: string;
   actorUuid: string | null;
   mpCost: number;
   judgeless: boolean;
-  /** 判定の成功数。judgeless のときは固定成功数 */
-  successCount: number | null;
-  damageTotal: number | null;
 };
 
 const defineKaiAttackCardSchema = () => {
@@ -21,27 +19,23 @@ const defineKaiAttackCardSchema = () => {
     actorUuid: new DocumentUUIDField({ type: "Actor", nullable: true, initial: null }),
     mpCost: new NumberField({ required: true, integer: true, min: 0, initial: 0 }),
     judgeless: new BooleanField({ required: true, initial: false }),
-    // 判定の成功数。judgeless のときは固定成功数が入る
-    successCount: new NumberField({ required: true, integer: true, nullable: true, initial: null }),
-    damageTotal: new NumberField({ required: true, integer: true, nullable: true, initial: null }),
   };
 };
 
 /**
  * 怪異の攻撃カードのChatMessage。
  *
- * 武器カードと違い育たない（判定とダメージを一度に振って1枚に出す）ので、状態は結果だけ持つ。
- * ボタンのハンドラ（ダメージ適用）は `data/` に置かず、`applications/` 側が持つものを
- * `emoklore.ts` の init が `ACTIONS` へ登録する。武器カードが `data/` でオーケストレータ化
- * している既知の課題（architecture.md 課題5）を、怪異カードでは繰り返さない。
+ * 1枚のカードが育つ形と、進み具合から決まるボタンの出し分けは `AttackCardModel` が持つ。
+ * ここに残るのは怪異の攻撃に固有の焼き込みだけになる。
+ *
+ * ボタンのハンドラは `data/` に置かず、`applications/` 側が持つものを `emoklore.ts` の
+ * init が `ACTIONS` へ登録する（どのカードも同じ形）。
  */
-export class KaiAttackCardModel extends ChatCardModel {
+export class KaiAttackCardModel extends AttackCardModel {
   declare attackName: string;
   declare actorUuid: string | null;
   declare mpCost: number;
   declare judgeless: boolean;
-  declare successCount: number | null;
-  declare damageTotal: number | null;
 
   static override CARD: CardIdentity = {
     root: ".em-kai-attack-card",
@@ -52,6 +46,11 @@ export class KaiAttackCardModel extends ChatCardModel {
   static override ACTIONS: CardActions<KaiAttackCardModel> = {};
 
   static override defineSchema() {
-    return defineKaiAttackCardSchema();
+    return { ...super.defineSchema(), ...defineKaiAttackCardSchema() };
+  }
+
+  /** 判定なしの攻撃は判定を振らないので、`message.rolls` の先頭がダメージになる */
+  protected override get hasAttackRoll(): boolean {
+    return !this.judgeless;
   }
 }
