+++ templates/chat/kai-attack-card.hbs
@@ -45,6 +45,10 @@
         <i class="fa-solid fa-heart-crack" inert></i>
         <span>{{localize "EMOKLORE.ChatMessage.attack.Apply"}}</span>
       </button>
+      <button type="button" data-action="applyDamageWithReduction">
+        <i class="fa-solid fa-shield-halved" inert></i>
+        <span>{{localize "EMOKLORE.ChatMessage.attack.ApplyWithReduction"}}</span>
+      </button>
     </footer>
   {{/if}}
 </div>
